refactor(pathutils): centralize repo name extraction and MCP paths
- Add extractRepoName to PathUtils with GitHub and Azure DevOps support
  and a sanitized fallback for arbitrary URLs
- Export getMcpPaths from PathUtils and remove duplicate implementations
- Update BackgroundProcessor and index to import and use shared helpers
- Adjust CLI metadata to accept generic Git repository URLs
- Minor whitespace and docs cleanup in PathUtils
