+++ src/core/PathUtils.ts
@@ -4,7 +4,7 @@ import fs from 'fs';
 /**
  * Utility functions for handling MCP folder paths with environment variable support
  * and platform-specific defaults
- * 
+ *
  * Note: This module is intentionally logging-free to avoid circular dependencies
  */
 
@@ -14,7 +14,7 @@ import fs from 'fs';
  */
 export function getDataFolder(): string {
   const envDataFolder = process.env.MCP_DATA_FOLDER;
-  
+
   if (envDataFolder) {
     return envDataFolder;
   }
@@ -28,7 +28,7 @@ export function getDataFolder(): string {
  */
 export function getDocsFolder(): string {
   const envDocsFolder = process.env.MCP_DOCS_FOLDER;
-  
+
   if (envDocsFolder) {
     return envDocsFolder;
   }
@@ -39,7 +39,7 @@ export function getDocsFolder(): string {
 /**
  * Get platform-specific default data folder
  * - Linux: ~/.local/share/local-search-mcp
- * - macOS: ~/Library/Application Support/local-search-mcp  
+ * - macOS: ~/Library/Application Support/local-search-mcp
  * - Windows: %LOCALAPPDATA%/local-search-mcp
  */
 function getDefaultDataFolder(): string {
@@ -49,11 +49,11 @@ function getDefaultDataFolder(): string {
   switch (platform) {
     case 'darwin': // macOS
       return path.join(homeDir, 'Library', 'Application Support', 'local-search-mcp');
-    
+
     case 'win32': // Windows
       const localAppData = process.env.LOCALAPPDATA || path.join(homeDir, 'AppData', 'Local');
       return path.join(localAppData, 'local-search-mcp');
-    
+
     default: // Linux and others
       return path.join(homeDir, '.local', 'share', 'local-search-mcp');
   }
@@ -68,6 +68,61 @@ function getDefaultDocsFolder(): string {
 }
 
 /**
+ * Get common MCP paths for use throughout the application
+ */
+export function getMcpPaths() {
+  const dataFolder = getDataFolder();
+  const docsFolder = getDocsFolder();
+
+  return {
+    data: dataFolder,
+    docs: docsFolder,
+    repositories: path.join(docsFolder, 'repositories'),
+    fetched: path.join(docsFolder, 'fetched'),
+    database: path.join(dataFolder, 'local-search-index.db'),
+    logs: path.join(dataFolder, 'local-search-mcp.log'),
+    temp: path.join(dataFolder, 'temp')
+  };
+}
+
+/**
+ * Extract repository name from URL (supports Azure DevOps and GitHub)
+ * Used for generating directory names and logging
+ */
+export function extractRepoName(repoUrl: string): string {
+  try {
+    const url = repoUrl.replace(/\.git$/, '');
+    // Check for Azure DevOps URLs
+    if (url.includes('dev.azure.com') || url.includes('visualstudio.com')) {
+      // Example: https://dev.azure.com/org/project/_git/repo
+      const parts = url.split('/').filter(p => p); // remove empty parts
+      const gitIndex = parts.findIndex(p => p === '_git');
+      if (gitIndex > 0 && parts.length > gitIndex + 1) {
+        const repo = parts[gitIndex + 1];
+        const project = parts[gitIndex - 1];
+        const org = parts[gitIndex - 2];
+        if (org && project && repo) {
+          return `${org}_${project}_${repo}`;
+        }
+      }
+    }
+    // Fallback to existing logic (GitHub style)
+    const parts = url.split('/');
+    if (parts.length >= 2) {
+      const owner = parts[parts.length - 2];
+      const repo = parts[parts.length - 1];
+      return `${owner}_${repo}`;
+    }
+
+    throw new Error('Invalid repo URL format');
+  } catch (error) {
+    // Generate a default name from the URL
+    const cleanUrl = repoUrl.replace(/^https?:\/\//, '').replace(/\.git$/, '');
+    return cleanUrl.replace(/[^a-zA-Z0-9]/g, '_').substring(0, 50) || `unknown_repo_${Date.now()}`;
+  }
+}
+
+/**
  * Ensure a directory exists, creating it recursively if needed
  * @param dirPath Directory path to create
  * @param description Description for debugging (not used in this logging-free version)
@@ -75,7 +130,7 @@ function getDefaultDocsFolder(): string {
 export async function ensureDirectoryExists(dirPath: string, description?: string): Promise<void> {
   try {
     await fs.promises.mkdir(dirPath, { recursive: true });
-    
+
     // Verify the directory was created and is accessible
     const stats = await fs.promises.stat(dirPath);
     if (!stats.isDirectory()) {
@@ -93,37 +148,19 @@ export async function initializeMcpDirectories(): Promise<void> {
   try {
     const dataFolder = getDataFolder();
     const docsFolder = getDocsFolder();
-    
+
     // Create base directories
     await ensureDirectoryExists(dataFolder, 'Data folder (database, logs)');
     await ensureDirectoryExists(docsFolder, 'Docs folder (documents)');
-    
+
     // Create docs subdirectories
     const repositoriesFolder = path.join(docsFolder, 'repositories');
     const fetchedFolder = path.join(docsFolder, 'fetched');
-    
+
     await ensureDirectoryExists(repositoriesFolder, 'Repositories folder');
     await ensureDirectoryExists(fetchedFolder, 'Fetched files folder');
-    
+
   } catch (error: any) {
     throw error;
   }
-}
-
-/**
- * Get common MCP paths for use throughout the application
- */
-export function getMcpPaths() {
-  const dataFolder = getDataFolder();
-  const docsFolder = getDocsFolder();
-  
-  return {
-    data: dataFolder,
-    docs: docsFolder,
-    repositories: path.join(docsFolder, 'repositories'),
-    fetched: path.join(docsFolder, 'fetched'),
-    database: path.join(dataFolder, 'local-search-index.db'),
-    logs: path.join(dataFolder, 'local-search-mcp.log'),
-    temp: path.join(dataFolder, 'temp')
-  };
 }
